ER Logical-To-Physical Implementation
Here we introduce an algorithm for performing logical-to-physical
transformations for ER models in the presence of path-equational
constraints.

Each constraint is represented as a pair of navigations. A navigation is
one or more relationships composed end-to-end. The diagonal is followed
from the source of the relationship. The riser is followed from the
target of the relationship. Both navigations must reach the same entity
type.

By following both the riser and the diagonal, we can produce a diagram.
The constraint asserts that this diagram is a commutativity diagram.

The algorithm supports zero constraints (whereupon the standard Chen
transformation is applied), one constraint (the Cartmell case) and more.

Included are test cases for some of the more interesting configurations.