Codegen: steps towards model bootstrapping
We are very close to being able to bootstrap our entity models, which is
to say describe our entity models in terms of an entity model. This meta
model is currently residing in `gen/test/entity_model`.

In order for this to be practical, we need a way of representing
sequences. A sequence is a dependency where the order of insertion
matters. This order may or may not be identifying. The sequence
implementation is currently quite sketchy. Parsing files works, that's
it. Some thought about how the CRUD operations interacts with sequences
is required.

More types of attributes are supported, rather than just supporting
strings. Very soon, the `rtl` package will be moved into the root, as
most of the root will become `meta`.

We are now vendoring the `top` and `unify` packages (the repositories of
which will be discontinued) and a go.mod file has been introduced.