Issue #3: Path-based L2P
Here is a very rough implementation of path-based relationships. A path
is a compact syntax for describing how a relationship may be
implemented.

This commit introduces the parser for path syntax, along with utilities
for manipulating them. The logical-to-physical program is now written
around this package, using path.Eval to drive the process. At its core,
L2P is analysing the types of the paths that it is shown, and filling in
any gaps that it finds.

The type of a path is represented here as a set of `source->destination`
mappings, where both `source` and `destination` are sets of entities. We
can talk about a set of entities here as being an entity type along with
the attributes on the type that can be proven to have a single value in
a particular context. Some examples follow.

Given our square model, the path `name` has type

    a(name) -> $() | b(name) -> $() | c(name) -> $() | d(name) -> $()

where `$()` represents the set of all attribute values.

The path `s_name/~name` has type

    a(s_name) -> a(name) | a(s_name) -> b(name) | a(s_name) -> c(name) |
    a(s_name) -> d(name)

The relationship `s` is defined in terms of this path, along with the
knowledge that the relationship is from `a` to `b`, for which we can use
the path `~/a//b&s_name/~name` with type

    a() -> b() | a(s_name) -> b(name)

Note here that the path is defined for when we have a single, known
instance of `b` and for when we have a set of `b`s.

The process of performing L2P is, therefore a matter of going through
each relationship, ensuring the identity of its target is fully known,
determining which attributes are known by following the relationship's
path, and creating any attributes that are missing and adding them to
the path of our relationship.